test-requirements.txt
black==25.1.0
build==1.2.2.post1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.10
defusedxml==0.7.1
flake8==7.1.1
flatten-dict==0.4.2
fuzzywuzzy==0.18.0
idna==3.10
iniconfig==2.0.0
langdetect==1.0.9
Levenshtein==0.27.1
lxml==5.3.1
mccabe==0.7.0
mock==5.1.0
mypy-extensions==1.0.0
mysql==0.0.3
mysql-connector-python==9.2.0
mysqlclient==2.2.7
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
pymarc==5.2.3
pyproject_hooks==1.2.0
pysolr==3.10.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-datadir==1.5.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
PyYAML==6.0.2
RapidFuzz==3.12.2
requests==2.32.3
setuptools==75.8.0
six==1.17.0
tomli==2.2.1
urllib3==2.3.0
wheel==0.45.1
xmltodict==0.14.2
